Move terminal to main tabs and make it default route
